+++ CHANGELOG.md
@@ -37,6 +37,7 @@
 * Registration module: check that input image does not contain multiple fields of view (axis F).
 * Fix handling of .ome.tif files.
 * Force non-editable napari layers to stay non-editable (In napari v0.4.17, changing an axis value using the corresponding slider makes the layer editable).
+* Set random seed in "feature matching" registration methods for reproducibility.